+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 * [FIX] Push notifications not being sent for TCP/IP connection errors
 * New device discovery
 * Add "Seperate Mode Accessories" setting
+* Add "Seperate Fan Zone Switches" setting
 
 ## 3.2.4 (2020-12-27)